backend/requirements.txt
fastapi
uvicorn
langchain
llama-index
chromadb
pypdf
docx2txt
python-dotenv
openai
anthropic
llama-cpp-python 